# src/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Idea Hub API
  description: 灵感管理系统的标准化 REST 接口规范
  version: 1.0.0

servers:
  - url: http://localhost:3000/api

paths:
  /ping:
    get:
      summary: 健康检查
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: "ok" }

  /ideas:
    get:
      summary: 获取idea列表（带分页）
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
          example: 1
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            default: 10
          example: 10
      responses:
        '200':
          description: 成功获取分页列表，可能为空
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedIdeas'
        '400':
          $ref: '#/components/responses/BadRequest'

    
    post:
      summary: 创建新灵感
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIdeaInput'
      responses:
        '201':
          description: Successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Idea'
        '400':
          $ref: '#/components/responses/BadRequest'

  /ideas/batch:
    get:
      summary: 批量获取灵感
      parameters:
        - name: ids
          in: query
          required: true
          schema:
            type: string
          example: "550e8400-e29b-41d4-a716-446655440000,550e8400-e29b-41d4-a716-446655440001"
      responses:
        '200':
          description: 成功获取批量灵感结果
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Idea'
        '400':
          $ref: '#/components/responses/BadRequest'

  /ideas/{id}:
    get:
      summary: 获取灵感详情
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Idea'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      summary: 部分更新灵感
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateIdeaInput'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Idea'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      summary: 删除灵感
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '400':
          $ref: '#/components/responses/BadRequest'
        '204':
          description: No Content (Successfully deleted)
        '404':
          $ref: '#/components/responses/NotFound'

components:
  schemas:
    IdeaStatus:
      type: string
      enum: [draft, published, archived]
      description: 灵感状态

    CreateIdeaInput:
      type: object
      additionalProperties: false
      required: [title]
      properties:
        title: {$ref: '#/components/schemas/IdeaTitle'}
        status: { $ref: '#/components/schemas/IdeaStatus', default: draft }

    UpdateIdeaInput:
      type: object
      additionalProperties: false
      description: 仅包含需要更新的字段
      properties:
        title: {$ref: '#/components/schemas/IdeaTitle'}
        status: { $ref: '#/components/schemas/IdeaStatus' }

    Idea:
      type: object
      additionalProperties: false
      required: [id, title, status, ownerId, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        title: { $ref: '#/components/schemas/IdeaTitle' }
        status: { $ref: '#/components/schemas/IdeaStatus' }
        ownerId:
          type: string
          format: uuid
          description: 服务端写入的资源归属用户 id，客户端不可在 request 中提交
        createdAt:
          type: string
          format: date-time
          description: 服务端生成的创建时间
        updatedAt:
          type: string
          format: date-time
          description: 服务端自动维护的最后更新时间
          
    IdeaTitle:
      type: string
      minLength: 2
      maxLength : 100
      example: '我的新项目'

    IdeaContent:
      type: string
      maxLength: 2000
      example: '这是一个深刻的创意...'
      
    PaginationMeta:
      type: object
      required: [total, page, pageSize, totalPages, hasNextPage]
      properties:
        total: { type: integer, example: 100 }
        page: { type: integer, example: 1 }
        pageSize: { type: integer, example: 10 }
        totalPages: { type: integer, example: 10 }
        hasNextPage: { type: boolean, example: true }

    PaginatedIdeas:
      type: object
      required: [items, meta]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Idea' }
        meta: { $ref: '#/components/schemas/PaginationMeta' }

    Error:
      type: object
      required: [code, message, requestId]
      properties:
        code: 
          type: string
          description: "业务错误码"
          example: "INVALID_PARAMS"
        message: 
          type: string
          description: "提示信息"
          example: "标题不能为空"
        requestId: 
          type: string
          description: "唯一请求ID"
          example: "req-123456"

  responses: 
    BadRequest: 
      description: "请求参数错误"
      content:
        application/json:
          schema: 
            $ref: '#/components/schemas/Error'
    NotFound:
      description: "资源未找到"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'